started work on task framework

+++ src/main/java/frc/lib/control/Subsystems/SimpleButtonHold.java
@@ -6,12 +6,12 @@ import frc.lib.output.Motors;
 public class SimpleButtonHold extends Subsystem {
     protected int button, motorID, controller;
     boolean reversed;
-    public SimpleButtonHold(long timing, boolean reversed, int button, int motorID){
+    public SimpleButtonHold(int timing, boolean reversed, int button, int motorID){
         super(timing);
         this.button = button;
         this.controller = 0;
     }
-    public SimpleButtonHold(long timing, boolean reversed, int button, int controller, int motorID){
+    public SimpleButtonHold(int timing, boolean reversed, int button, int controller, int motorID){
         super(timing);
         this.button = button;
         this.controller = controller;
